+++ .gulp/misc.js
@@ -1,37 +1,56 @@
 import gulp from 'gulp';
-const { src, dest } = gulp
+import connect from 'gulp-connect';
+import plumber from 'gulp-plumber';
 
-import connect from 'gulp-connect'
+/**
+ * Custom
+ */
+import { errorHandler } from './lib/utils.js';
 
-const copyFiles = () =>
-  src(
-    [
-      './src/_redirects',
-      './src/_headers',
-      './src/robots.txt',
-      './src/favicon.ico',
-    ],
-    {
-      allowEmpty: true,
-    }
-  )
-    .pipe(dest('./dist/'))
+const { src, dest } = gulp;
+
+const copyTest = () =>
+  src('./src/test/**/*')
+    .pipe(plumber({ errorHandler }))
+    .pipe(dest('./dist/test/'))
     .pipe(connect.reload())
 
+
+const copyFiles = () =>
+  src([
+    './src/_redirects',
+    './src/_headers',
+    './src/robots.txt',
+    './src/favicon.ico',
+  ])
+    .pipe(plumber({ errorHandler }))
+    .pipe(dest('./dist/'))
+
+
 const copyBuildFiles = () =>
-  src(['./dist/**/*'], {
-    allowEmpty: true,
-  }).pipe(dest('./build/'))
+  src(['./dist/**/*'])
+    .pipe(plumber({ errorHandler }))
+    .pipe(dest('./build/'))
+
 
 const copyVideo = () =>
-  src('./src/video/**/*').pipe(dest('./dist/video/')).pipe(connect.reload())
+  src('./src/video/**/*')
+    .pipe(plumber({ errorHandler }))
+    .pipe(dest('./dist/video/'))
+    .pipe(connect.reload())
+
 
 const copyFonts = () =>
-  src('./src/fonts/**/*').pipe(dest('./dist/fonts/')).pipe(connect.reload())
+  src('./src/fonts/**/*')
+    .pipe(plumber({ errorHandler }))
+    .pipe(dest('./dist/fonts/'))
+    .pipe(connect.reload())
+
 
 const copyIcons = () =>
   src('./src/favicons/**/*')
+    .pipe(plumber({ errorHandler }))
     .pipe(dest('./dist/favicons/'))
     .pipe(connect.reload())
 
-export { copyFiles, copyBuildFiles, copyVideo, copyFonts, copyIcons }
+export { copyTest, copyFiles, copyBuildFiles, copyVideo, copyFonts, copyIcons };