+++ .gulp/styles.js
@@ -1,3 +1,8 @@
+/**
+ * Styles
+ * ================================================================================
+ */
+
 import gulp from 'gulp';
 const { src, dest } = gulp;
 
@@ -11,13 +16,12 @@ import connect from 'gulp-connect';
  * Notification
  */
 import plumber from 'gulp-plumber';
-import notify from 'gulp-notify';
 import size from 'gulp-size';
 
 /**
  * Styles
  */
-import dartSass from 'sass';
+import * as dartSass from 'sass';
 import gulpSass from 'gulp-sass';
 const sass = gulpSass(dartSass);
 
@@ -34,10 +38,15 @@ import cssnano from 'cssnano';
  */
 import sourcemaps from 'gulp-sourcemaps';
 
+/**
+ * Custom
+ */
+import { errorHandler } from './lib/utils.js';
+
 
 /**
-* Config
-*/
+ * Config
+ */
 const cfg = {
   src: {
     scss: './src/scss/**/*.scss',
@@ -51,39 +60,42 @@ const cfg = {
 
 /**
  * Styles
- * ================================================================================
+ *
+ * @description Compile SCSS to CSS
  */
-const scss = () =>
+const scss = (done) =>
   src(cfg.src.scss)
-    .pipe(plumber())
+    .pipe(plumber({ errorHandler }))
     .pipe(sourcemaps.init())
     .pipe(
       sass({
-        outputStyle: 'expanded',
+        style: 'expanded',
         errLogToConsole: false,
-        includePaths: ['node_modules', 'bower_components', 'src', '.'],
+        loadPaths: ['node_modules', 'bower_components', 'src', '.'],
+        silenceDeprecations: ['import'],
+        quietDeps: true,
       })
     )
-    .on('error', notify.onError())
     .pipe(sourcemaps.write('./'))
     .pipe(dest(cfg.dest.scss))
-    .pipe(dest(cfg.dest.css));
+    .pipe(dest(cfg.dest.css))
+    .on('end', done);
 
 /**
  * Styles Reload
- *
  */
-const stylesReload = () => src(cfg.dest.scss)
-  .pipe(connect.reload());
+const stylesReload = (done) =>
+  src(cfg.dest.scss)
+    .pipe(connect.reload())
+    .on('end', done);
 
 /**
-* PostCSS, Autoprefixer, CSS compressor
-*/
+ * PostCSS, Autoprefixer, CSS compressor
+ */
 const cssCompress = (done) =>
   src('./dist/styles/**/*.css')
-    .pipe(plumber())
+    .pipe(plumber({ errorHandler }))
     .pipe(postcss([autoprefixer(), cssnano()]))
-    .on('error', notify.onError())
     .pipe(dest('./build/styles/'))
     .pipe(size())
     .pipe(connect.reload())