+++ .gulp/tests.js
@@ -1,8 +1,8 @@
 // Tests
 
-import mobileFriendlyTest from 'mobile-friendly-test-npm'
-import htmlSpeed from 'html-speed'
-import cssTest from 'css-test-npm'
+import mobileFriendlyTest from 'mobile-friendly-test-npm';
+import htmlSpeed from 'html-speed';
+import cssTest from 'css-test-npm';
 
 import dotenv from 'dotenv';
 dotenv.config();
@@ -12,8 +12,15 @@ dotenv.config();
  */
 const mftApiKey = process.env.MFT_KEY || '';
 const mftUrl = process.env.PROXY_URL || '';
-const mobileTestRes = () =>
+
+const mobileTestRes = (done) => {
   mobileFriendlyTest(mftUrl, mftApiKey)
+    .then(() => done())
+    .catch((error) => {
+      console.error(error);
+      done(error);
+    });
+};
 
 
 /**
@@ -21,14 +28,29 @@ const mobileTestRes = () =>
  */
 const hstApiKey = process.env.HST_KEY || '';
 const hstUrl = process.env.PROXY_URL || '';
-const htmlSpeedRes = () =>
+
+const htmlSpeedRes = (done) => {
   htmlSpeed(hstUrl, hstApiKey)
+    .then(() => done())
+    .catch((error) => {
+      console.error(error);
+      done(error);
+    });
+};
 
 
 /**
  * CSS Test
  */
 const cssUrl = process.env.PROXY_URL || '';
-const cssTestRes = () => cssTest(cssUrl)
 
-export { mobileTestRes, htmlSpeedRes, cssTestRes }
+const cssTestRes = (done) => {
+  cssTest(cssUrl)
+    .then(() => done())
+    .catch((error) => {
+      console.error(error);
+      done(error);
+    });
+};
+
+export { mobileTestRes, htmlSpeedRes, cssTestRes };