+++ src/javascript/app.js
@@ -9,7 +9,7 @@ import ModuleClass from 'src/module-class';
 import Component from 'src/components/component';
 import ResourcesLoaderTest from 'src/resources-loader-test.js';
 
-import 'ui-explorer';
+// import 'ui-explorer';
 
 // Check JS is enabled
 NoJs();